refactor: split up GeneralStrategy component
